requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
Django
django-cors-headers
django-environ
djangorestframework==3.15.2
gunicorn==23.0.0
markdownify==0.14.1
packaging==24.2
psycopg2==2.9.10
PyJWT
python-dotenv==1.0.1
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise
